// src/generated/sensor/general_sensor_data.ts
export enum SensorName {
  APRIL_TAGS = 0,
  ODOMETRY = 1,
  IMU = 2,
  LIDAR = 3,
  CAMERA = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface GeneralSensorData {
  sensorName: SensorName;
  sensorId: string;
  timestamp: string;
  processingTimeMs: string;
  apriltags?: AprilTagData | undefined;
  odometry?: OdometryData | undefined;
  image?: ImageData | undefined;
  lidar?: LidarData | undefined;
  imu?: ImuData | undefined;
}